requirements.txt
appnope==0.1.4
asttokens==2.4.1
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
executing==2.1.0
filelock==3.16.1
identify==2.6.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
